Antigravity-AgenticAIs copy/requirements.txt
# Project Aegis - Requirements
# Install with: pip install -r requirements.txt

# Core LangChain and LangGraph
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.20

# Google Generative AI
google-generativeai>=0.3.0

# Quantum Computing (for Grover's algorithm)
qiskit>=1.0.0
qiskit-aer>=0.13.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
matplotlib>=3.5.0

# Database (for future MySQL integration)
# Uncomment when ready to use MySQL:
# mysql-connector-python>=8.0.0
# sqlalchemy>=2.0.0
